Major bugfix

+++ repulsion.h
@@ -24,8 +24,8 @@ double B_term(const int i1, const int i2, const int r1, const int r2, const int 
 		const double px, const double ax, const double bx, const double qx, const double cx, const double dx, const double gamma1,
 		const double gamma2, const double delta);
 
-double fB(const int i, const int l1, const int l2, const int p, const int a, const int b, const int r, const int q);
-double B0(int i, int r, int q);
+double fB(const int i, const int l1, const int l2, const double p, const double a, const double b, const int r, const double q);
+double B0(int i, int r, double q);
 double fact_ratio2(const int a, const int b);
 
 const unsigned int teindex(unsigned int i, unsigned int j, unsigned int k, unsigned int l);
